mode customChainIdsToAssetNames
