+++ src/core/references/chains.ts
@@ -1,43 +1,8 @@
 import backendChains from 'static/data/chains.json';
-import {
-  Chain,
-  arbitrumNova,
-  aurora,
-  avalanche,
-  blast,
-  blastSepolia,
-  boba,
-  canto,
-  celo,
-  classic,
-  cronos,
-  degen,
-  dogechain,
-  fantom,
-  filecoin,
-  harmonyOne,
-  immutableZkEvm,
-  kava,
-  klaytn,
-  linea,
-  manta,
-  mantle,
-  metis,
-  mode,
-  moonbeam,
-  opBNB,
-  palm,
-  pgn,
-  polygonZkEvm,
-  pulsechain,
-  rootstock,
-  scroll,
-  zkSync,
-} from 'viem/chains';
+import { Chain } from 'viem/chains';
 
 import {
   BackendNetwork,
-  ChainId,
   ChainName,
   chainHardhat,
   chainHardhatOptimism,
@@ -97,46 +62,3 @@ export const meteorologySupportedChains = backendChains.networks
 export const needsL1SecurityFeeChains = backendChains.networks
   .filter((backendChain: BackendNetwork) => backendChain.opStack)
   .map((backendChain: BackendNetwork) => parseInt(backendChain.id, 10));
-
-export const customChainIdsToAssetNames: Record<ChainId, string> = {
-  [arbitrumNova.id]: 'arbitrumnova',
-  [aurora.id]: 'aurora',
-  [avalanche.id]: 'avalanchex',
-  [blast.id]: 'blast',
-  [blastSepolia.id]: 'blastsepolia',
-  [boba.id]: 'boba',
-  [celo.id]: 'celo',
-  [classic.id]: 'classic',
-  [cronos.id]: 'cronos',
-  [degen.id]: 'degen',
-  [dogechain.id]: 'dogechain',
-  [fantom.id]: 'fantom',
-  [filecoin.id]: 'filecoin',
-  [harmonyOne.id]: 'harmony',
-  [immutableZkEvm.id]: 'immutablezkevm',
-  [kava.id]: 'kavaevm',
-  [klaytn.id]: 'klaytn',
-  [linea.id]: 'linea',
-  957: 'lyra',
-  [manta.id]: 'manta',
-  [mantle.id]: 'mantle',
-  [metis.id]: 'metis',
-  [mode.id]: 'mode',
-  [moonbeam.id]: 'moonbeam',
-  [canto.id]: 'nativecanto',
-  [opBNB.id]: 'opbnb',
-  [palm.id]: 'palm',
-  [pgn.id]: 'pgn',
-  [polygonZkEvm.id]: 'polygonzkevm',
-  [pulsechain.id]: 'pulsechain',
-  1380012617: 'rari',
-  1918988905: 'raritestnet',
-  690: 'redstone',
-  17069: 'redstonegarnet',
-  17001: 'redstoneholesky',
-  [rootstock.id]: 'rootstock',
-  31: 'rootstocktestnet',
-  [scroll.id]: 'scroll',
-  100: 'xdai',
-  [zkSync.id]: 'zksync',
-};
